Actualización: Agregada función general de cálculo (en progreso)

+++ src/funciones/Trigonometria.java
@@ -3,15 +3,15 @@ package funciones;
 import java.math.BigDecimal;
 
 public class Trigonometria {
-    public static double sen(BigDecimal x) {
-        return Math.sin(x.doubleValue());
+    public static BigDecimal sen(BigDecimal x) {
+        return BigDecimal.valueOf(Math.sin(x.doubleValue()));
     }
 
-    public static double cos(BigDecimal x) {
-        return Math.cos(x.doubleValue());
+    public static BigDecimal cos(BigDecimal x) {
+        return BigDecimal.valueOf(Math.cos(x.doubleValue()));
     }
 
-    public static double tan(BigDecimal x) {
-        return Math.tan(x.doubleValue());
+    public static BigDecimal tan(BigDecimal x) {
+        return BigDecimal.valueOf(Math.tan(x.doubleValue()));
     }
 }
